refactor: contextprocessor in its own file, use processedattachments for form data
