+++ src/components/App.tsx
@@ -9,8 +9,6 @@ import { useStore } from '@/providers/store-provider'
 import Modals from './modals/Modals'
 import { ModalContainer } from '@/components/modals/ModalContainer'
 import { useShallow } from 'zustand/react/shallow'
-import { dataURItoFile } from '@/utils/attachments'
-import { Attachment, ImageAttachment, PdfAttachment, SpreadsheetAttachment, TextFileAttachment } from '@/types'
 import { initAmplitude, trackEvent } from '@/utils/amplitude'
 import useAuth from '@/hooks/useAuth'
 import { decodeJwt } from 'jose'
@@ -19,60 +17,7 @@ import ToastContainer from '@/components/Toast/ToastContainer'
 import usePromptApps from '@/hooks/usePromptApps'
 import { useChatHistory } from '@/hooks/useChatHistory'
 import { Prompt } from '@/types/supabase-helpers'
-
-function processAttachments(attachments: any[]): Attachment[] {
-  return attachments.map((attachment) => {
-    if (attachment.type !== 'file') {
-      return attachment as Attachment
-    }
-    const { fileName, mimeType, value } = attachment
-
-    if (mimeType.startsWith('image/')) {
-      return {
-        type: 'image',
-        value,
-      } as ImageAttachment
-    } else if (mimeType === 'application/pdf') {
-      console.log('Processing PDF:', fileName, mimeType, value)
-      const file = dataURItoFile(attachment.value, fileName, mimeType)
-      if (!file) {
-        console.error('Could not convert data URI to file for PDF:', fileName, value)
-        return attachment
-      }
-      console.log('Converted PDF to file:', file, file.size, file.type, file.name)
-      return {
-        type: 'pdf',
-        value: file,
-      } as PdfAttachment
-    } else if (
-      mimeType.includes('spreadsheetml') ||
-      mimeType.includes('excel') ||
-      attachment.fileName.endsWith('.xlsx')
-    ) {
-      return {
-        type: 'spreadsheet',
-        value,
-      } as SpreadsheetAttachment
-    } else if (
-      mimeType.startsWith('text/') ||
-      mimeType === 'application/json' ||
-      mimeType === 'application/xml' ||
-      mimeType === 'application/javascript' ||
-      mimeType === 'application/vnd.openxmlformats-officedocument.wordprocessingml.document' ||
-      mimeType === 'application/msword' ||
-      mimeType === 'application/vnd.openxmlformats-officedocument.presentationml.presentation'
-    ) {
-      return {
-        type: 'text_file',
-        value,
-        fileName,
-      } as TextFileAttachment
-    } else {
-      console.error('Unprocessable attachment:', mimeType, fileName, value)
-      return attachment
-    }
-  })
-}
+import { processAttachments } from '@/utils/contextprocessor'
 
 function useContextReceivedHandler(navigateToNewChat: () => void) {
   const {
